+++ node/tests/TestUtilities.ts
@@ -392,6 +392,8 @@ export async function transactionTest(
     args.push(new Set(["bar", "foo"]));
     baseTransaction.sinter([key7, key7]);
     args.push(new Set(["bar", "foo"]));
+    baseTransaction.sinterstore(key7, [key7, key7]);
+    args.push(2);
     baseTransaction.sdiff([key7, key7]);
     args.push(new Set());
     baseTransaction.srem(key7, ["foo"]);